feat: add Certificatable protocol for any object
